fixed error and added lookup argument for assignment
error fixed : if value instead of is value is not none that returned none when the value was 0.

lookup is used in assignment : when it's true, it tries to overwrite an existing variable at higher scope (default behavior, it's what happens in all for statements). when it's false, it does not try to do so (which allows for temporary variable such as for loops iterator variables)